selfai-backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
aiosmtplib==3.0.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asyncpg==0.30.0
attrs==25.1.0
Authlib==1.5.2
bcrypt==4.2.1
beautifulsoup4==4.13.3
billiard==4.2.1
blinker==1.9.0
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.1
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
environs==14.1.1
fastapi==0.115.8
fastapi-mail==1.4.2
fastapi_cors==0.0.6
frozenlist==1.5.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.67.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.4.2
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.51
langchain-ollama==0.2.3
langchain-openai==0.3.12
langchain-text-splitters==0.3.6
langsmith==0.3.8
lxml==5.3.1
lz4==4.4.3
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mistune==3.1.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
ollama==0.4.7
openai==1.73.0
orjson==3.10.15
outcome==1.3.0.post0
packaging==24.2
passlib==1.7.4
pdfminer.six==20240706
pillow==11.1.0
pinecone==6.0.1
pinecone-plugin-interface==0.0.7
pluggy==1.5.0
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
protoc-gen-openapiv2==0.0.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.3
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
qrcode==8.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
selenium==4.28.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.38
sqlmodel==0.0.22
starlette==0.45.3
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.18.3
zstandard==0.23.0
